scientific literature
Suppose that rather than wanting to estimate the spread or the proportion of blue, [][I'm interested only in the question, are there more blue beads than red beads]?  Another way to ask that is, *is 2p minus 1 bigger than 0*?

Is the spread bigger than 0?  So [][suppose we take a random sample of, say, 100 beads, and we observe 52 blue beads].  This gives us a spread of 4% {52/100 - (1 - 52/100) = 104/100 - 100/100 = 4/100}.  This seems to be pointing to there being more blue beads than red beads, because 4% is larger than 0.  52% is larger than 48%.  {but we all know its just one sample}

{Here, think about why we introduced the P value and null hypothesis?   We dont know P, we observed X_bar}

And you need to construct null hypothesis based on different condition.  Think, Think, Think

And 1 - (pnorm(z) - pnorm(Z)), appearing the graph in your head, and think

*In this case, there's actually a large chance of seeing 52 blue beads or more under the null hypothesis that there is the same amount of blue beads as red beads.  So the 52 blue beads are not very strong evidence, are not very convincing, if we want to make the case that there's more blue beads than red beads*.